NAS-137419 / 26.04 / Fix NFS exports mako (#17125)
By design, we disable ZFS `sharenfs` and make `/etc/exports.d`
immutable. We expect the directory to be empty.
Before starting NFS we confirm the directory is empty. If it is not we
look for `zfs.exports` and process it's contents to disable `sharenfs`
for that dataset. We also delete the file and all other files in the
directory and finally restore immutability to the directory.

The problem was that in a previous PR #17093 I erroneously widened the
file test to include 'zfs.exports.lock`. This caused an exception which
resulted in blocking NFS start.

### The Fix:
Limit the additional zfs processing to `zfs.exports`.   

This particular problem is Halfmoon specific.

See CI test console output
[here](http://jenkins.eng.ixsystems.net:8080/job/tests/job/api_tests/5800/console)