NAS-137424 / 26.04 / call failover.become_passive in system.reboot on HA (#17126)
We've had 2 or 3 rare situations where an HA customer will use the UI to
upgrade their HA system. The UI issues the `system.reboot` API call.
This works in 99.99% of all scenarios. But there are a few, incredibly
hard to reproduce, scenarios where it's not working.

The "proper" shutdown process of linux is untenable for how we've
written our failover logic so instead of trying to add even more
fragile, unreliable code, we'll take the big hammer approach and
"forcefully" ( 😉 ) reboot the active node.

NOTE: this only occurs when it's HA and HA has NOT been administratively
disabled. This is a very targeted change to keep regression
possibilities as low as possible.