NAS-137424 / 25.04.2.4 / call failover.become_passive in system.reboot on HA (by yocalebo) (#17127)
Automatic cherry-pick failed. Please resolve conflicts by running:

    git reset --hard HEAD~1
    git cherry-pick -x bdb3fbf3057ed08330b7260ea2c3ccb68b516e0f

If the original PR was merged via a squash, you can just cherry-pick the
squashed commit:

    git reset --hard HEAD~1
    git cherry-pick -x bab41016e625449c9477842dff91c8b91ec0909f

We've had 2 or 3 rare situations where an HA customer will use the UI to
upgrade their HA system. The UI issues the `system.reboot` API call.
This works in 99.99% of all scenarios. But there are a few, incredibly
hard to reproduce, scenarios where it's not working.

The "proper" shutdown process of linux is untenable for how we've
written our failover logic so instead of trying to add even more
fragile, unreliable code, we'll take the big hammer approach and
"forcefully" ( 😉 ) reboot the active node.

NOTE: this only occurs when it's HA and HA has NOT been administratively
disabled. This is a very targeted change to keep regression
possibilities as low as possible.

Original PR: https://github.com/truenas/middleware/pull/17126

Co-authored-by: Caleb St. John <30729806+yocalebo@users.noreply.github.com>