NAS-137334 / 25.10-RC.1 / add vm.device.convert (#17104)
Add VM disk image conversion functionality. This is a replacement (and
an improvement) from a previous
[commit](https://github.com/truenas/middleware/pull/13701) that
attempted to add similar functionality (but was never utilized). This
replaces the separate import/export operations with a unified convert
method that supports bidirectional conversion between disk images and
ZFS volumes. We automatically detect the direction of the conversion
based on the source/destination having the file extension.