NAS-137334 / 26.04 / add vm.device.convert (by yocalebo) (#17133)
Add VM disk image conversion functionality. This is a replacement (and
an improvement) from a previous
[commit](https://github.com/truenas/middleware/pull/13701) that
attempted to add similar functionality (but was never utilized). This
replaces the separate import/export operations with a unified convert
method that supports bidirectional conversion between disk images and
ZFS volumes. We automatically detect the direction of the conversion
based on the source/destination having the file extension.

Original PR: https://github.com/truenas/middleware/pull/17104

---------

Co-authored-by: Caleb St. John <30729806+yocalebo@users.noreply.github.com>
Co-authored-by: caleb <yocalebo@gmail.com>