NAS-137434 / 25.10-RC.1 / Fix disk temps when empty SD card readers are present (by yocalebo) (#17135)
In previous commit 294a283408bfa4ba9a01b845bef83ab15cb81f06, the ability
to create zpools when an EMPTY SD card reader was present was resolved.
However, the underlying issue is still present in other parts of the
stack because we also use this for gathering temperature information.

The issue is that in the `identifier` property of `DiskEntry` class goes
through a process to try and determine a unique identifier for every
disk on a system. This is important for obvious reasons. One of the
steps for determining the unique identifier is to read the disk for
partitions since we use GPT (they have GUID's). What's happening,
however, is that empty SD card readers aren't reporting a serial number
so we fallback to reading partitions. However, when the SD card reader
is empty the `open` syscall is failing with errno 123 (no medium found)
and we're not expecting the crash.

To remedy this situation we'll check for this errno specifically and log
all others. It's important that we don't crash here since this
functionality is tied to various other parts of the system in general
and crashing here causes a cascading set of failures.

Original PR: https://github.com/truenas/middleware/pull/17134

---------

Co-authored-by: caleb <yocalebo@gmail.com>