NAS-137054 / 26.04 / Improve defaults and validation for secure boot in VMs (#17107)
## Problem

When secure boot is enabled for a VM, we are defaulting for
`OVMF_CODE.fd` which won't work for VMs who want secure boot enabled.
This causes problems and for VMs which are already installed, changing
`bootloader_ovmf` can actually be problematic and result in an erroneous
configuration because of how nvram files for VMs are generated.

## Solution

Validation has been added to ensure that when secure boot is enabled for
a VM, we do not allow configuring a `bootloader_ovmf` file which will
not work with secure boot and secondly on vm update operation, we should
not allow changing `bootloader_ovmf` as that can lead to the VM not
booting up at all because of firmware issues.

Also when a VM is being created, we should default to a bootloader ovmf
file which reflects the secure boot setting properly.