NAS-137054 / 25.10-RC.1 / Improve defaults and validation for secure boot in VMs (by Qubad786) (#17138)
Automatic cherry-pick failed. Please resolve conflicts by running:

    git cherry-pick -x 57f14db90764c21af59aeb09dc7e0cef8b0051e3

If the original PR was merged via a squash, you can just cherry-pick the
squashed commit:

    git reset --hard HEAD~1
    git cherry-pick -x 72c646b7cdf187563eba2dc08b88c06da1eaaa45

## Problem

When secure boot is enabled for a VM, we are defaulting for
`OVMF_CODE.fd` which won't work for VMs who want secure boot enabled.
This causes problems and for VMs which are already installed, changing
`bootloader_ovmf` can actually be problematic and result in an erroneous
configuration because of how nvram files for VMs are generated.

## Solution

Validation has been added to ensure that when secure boot is enabled for
a VM, we do not allow configuring a `bootloader_ovmf` file which will
not work with secure boot and secondly on vm update operation, we should
not allow changing `bootloader_ovmf` as that can lead to the VM not
booting up at all because of firmware issues.

Also when a VM is being created, we should default to a bootloader ovmf
file which reflects the secure boot setting properly.

Original PR: https://github.com/truenas/middleware/pull/17107

Co-authored-by: M. Rehan <mrehanlm93@gmail.com>