NAS-137436 / 25.10-RC.1 / improve validation in vm.device.convert (by yocalebo) (#17141)
This logic was mistakingly removed in one of my force commits when I
originally added this endpoint. This presents a better error message
when neither the source nor destination fields in the API represent the
disk image to be converted from (or to).

Original PR: https://github.com/truenas/middleware/pull/17139

Co-authored-by: caleb <yocalebo@gmail.com>