Simplify ACL handling in samba for POSIX ACLs
This commit removes the smb.conf generation step in which we
add vfs_acl_xattr if POSIX ACLs are enabled. This was originally
added to provide support for a somewhat richer ACL model when
POSIX acltype was selected for an SMB share (non-default configuration).
Unfortunately, this creates two potential sources of truth for
the ACL contents and has potential to create hard-to-diagnose
permissions issues on the backend.

Since we already have a rich ACL model provided by the NFSv4 acltype on
ZFS, it provides a better UX to simplify the ACL implementation for
POSIX ACLs by removing this module.

(cherry picked from commit ffd0c4174fa7ebf966ab62c74239bec63390c4d9)