NAS-136978 / 26.04 / Fix link aggregation (#17124)
## Problem

The bond (LAG) interface is not being configured correctly, which causes
issues during failover. Currently, the `miimon` parameter is disabled.
As a result, the bond interface cannot switch to a working interface if
the active slave fails. The `miimon` parameter continuously monitors the
state of inactive interfaces and triggers a switch when the primary
interface goes down.

## Solution

Set the `miimon` value to `100`, as recommended in the [Linux kernel
bonding
documentation](https://www.kernel.org/doc/html/latest/networking/bonding.html),
to ensure proper bond interface handling.

It is essential to have this or `arp_interval` params set to have link
monitoring in place.