NAS-137491 / 26.04 / use new libzfs module in pool.dataset.update (#17166)
Use new pylibzfs module in pool.dataset.create. API stays the same,
tests pass. Took the liberty to create a namedtuple of the properties
that we iterate over in this method. It was getting a bit unwieldy since
each entry had 4 attributes and I ended up having a 5th. A named tuple
allows us to name each field so it's a bit more understandable.