NAS-137493 / 26.04 / Add global spotlight support for SMB (#17170)
This commit adds support for enabling the spotlight RPC server on samba
to talk to our truenas search server. It is enabled globally for all
shares in order to simplify UX.

The commit presupposes that the truenas search server is properly
configured and running. We can add more validation and tests once the
search / index daemon has been reviewed and merged into the product.

Search protocols is a list since eventually we'll add WSP support.