NAS-137549 / 25.10.0 / Allow workgroups named "DOMAIN" in SMB configuration (by anodos325) (#17200)
This commit relaxes validation to allow a customer to use a microsoft
reserved keyword for the workgroup / AD domain although this technically
should not be used. This is because re-deploying an AD domain to fix a
violation of MS
naming rules is onerous.

Original PR: https://github.com/truenas/middleware/pull/17195

Co-authored-by: Andrew Walker <andrew.walker@truenas.com>