NAS-137593 / 26.04 / Fix `ipmi.lan.update` (by creatorcary) (#17223)
Using `ipaddress.IPv4Address` as the field type when `strict=True` in
the `BaseModel` config expects a Python object to be passed, not a
string. We have custom types for this reason. This is the only use of
`ipaddress` types in our API schema, so no more changes are needed.

Original PR: https://github.com/truenas/middleware/pull/17221

---------

Co-authored-by: Logan Cary <logan.cary@ixsystems.com>