NAS-137580 / 26.04 / Syslog-ng is not respecting vendor preset file (#17227)
This PR adds changes to make sure that syslog-ng is enabled after the
system boots. Currently in trixie it is not respecting the preset file
we ship for some reason, hence we explicitly enable it to make sure it
is enabled when the system starts as lots of things rely on it to
function properly.