NAS-137576 / 26.04 / Fix netdata reporting (#17229)
## Problem

Most of our reporting plugins are not working right now due to trixie
upgrade which resulted in newer version of netdata being pulled in which
expects to operate differenlty.

## Solution

Quite a few of the netdata plugins work just fine after we update the
new plugins path and also go plugin has been enabled so we can get ups
reporting metrics as well as the older shell implemetnation we had
relied on upstream netdata ups plugin which is no longer there.