NAS-137684 / 26.04 / Fix containers naming regex to support backward compatibility (#17237)
## Problem

For incus containers we support `-` as valid naming character too and in
newer container implementation we don't. Which means that migration from
older implemetnation to newer one breaks if we find a container
containing a hyphen.

## Solution

Allow `-` in container names as well which in turn means migration from
incus to new containers works just fine. Alternatively if we don't want
to support `-` in container names, we can discuss how we want to rename
such incus containers.