NAS-137697 / 26.04 / Fix crash on unpicklable error (#17245)
In the production environment, this code is never reached.

However, in integration tests environment we pickle raised exceptions to
re-raise them on the test client.

`libvirt` errors are unpicklable (they are created by C code). When we
tried to send a pickled `libvirt` error through the wire, unhandled
`pickle.PickleError` was crashing JSON-RPC 2.0 server.