NAS-137730 / 26.04 / Remove deprecated setup.py install (#17258)
This shifts the middlewared package into using the modern approach for
building. There is a limitation with pyproject.toml in that you can only
have one file for one package. There isn't a way to have a single
pyproject.toml and have two modules. This means the
`src/middlewared/middlewared/test` subdirectory will be shipped by
default. There is no real harm in doing that anyways (this doesn't make
the package balloon in size) and it greatly simplifies how we run our
CI.

NOTE: We do not ship the test dependencies, only the test files.