NAS-137758 / 25.10.0 / Fix issue when LUN is added to an ALUA target (by bmeagherix) (#17277)
Perform a rescan of the internal HA target when a LUN is added to an
existing target (when ALUA is already enabled).

After the Trixie rebase the test `test_alua_luns` (and more) started
[failing](http://jenkins.eng.ixsystems.net:8080/job/tests/job/sharing_protocols_tests/518/)
reliably. The rebase appears to have highlighted a shortcoming/fragility
in the code.

----
iSCSI tests passing CI
[here](http://jenkins.eng.ixsystems.net:8080/job/tests/job/sharing_protocols_tests/521/).

Original PR: https://github.com/truenas/middleware/pull/17273

Co-authored-by: Brian M <brian.meagher@ixsystems.com>