Fix access check for group@ / GROUP_OBJ entries
This commit fixes access checks for group@ / GROUP_OBJ ACL entries.
There was a typo that crept in while refactoring this for new API
that resulted in the access for the owning group being evaluated as
if it were the user. This usually didn't manifest as an error
because the primary heavy ACL users are in active directory environments
where the NSS backend implements ID_TYPE_BOTH (so there's an identical
user and group entry). In the case of local accounts though it could
result in spurious validation errors (although in the ticket this is
associated with the user actually was trying to set permissions that
would raise a validation error even with the fix).

(cherry picked from commit f389cd6ceae4c6f836903b951c653fd4f2e4c761)