NAS-137817 / 26.04 / STIG: add nginx auditing (#17302)
In STIG mode we need to audit nginx.

This PR:
Enables auditing of nginx configuration files and logs in STIG mode.
Adds a CI end-to-end test to test a sample of nginx escalation events
and a general STIG privileged event.

This PR is dependent on the same named branch from truenas/audit_rules.

Successfully ran the CI tests locally.