NAS-137844 / 26.04 / fix USB disk detection alert on HA (#17308)
Raised by the support team. We need to ignore USB CDROM devices in this
alert so false positive proactive support tickets are not created in
support portal. I also switched to using `os.scandir` since this is
being backported to GE which uses python 3.11 and `pathlib.Path.iterdir`
on that python version is not performant compared to `os.scandir`.