NAS-137891 / 26.04 / remove last remnants of zfs.dataset.update (#17322)
This removes a few more calls to `zfs.dataset.update` since those crept
in with recent container API changes. This converts the data class to a
TypedDict because we made calls to `zfs.dataset.update` from a few API
tests. This fixes the tests and removes all the call sites. Tests pass.