NAS-137939 / 26.04 / Restore truenas_admin home directory ownership (#17342)
In the Source stanza of Debian control modules there is a setting:
`Rules-Requires-Root`.
If the setting is `binary-targets`, then existing file ownership is
retained when packaged. If the setting is `no`, then file ownership is
changed to `root` when packaged.

Pre-Trixie the default setting for `Rules-Requires-Root` was
`binary-targets`. In Trixie the default changed to `no`.

The truenas_admin home directory is part of the truenas-files Debian
package. It has no `Rules-Requires-Root` setting which allowed it to
follow the default setting that changed in Trixie. This resulted in the
ownership for the truenas_admin home directory getting changed to
`root`.

-------------------------------------------------

### This PR
Restores the pre-Trixie behavior with an explicit `Rules-Requires-Root:
binary-targets` setting.

Also in this PR:
- Removed a cp statement that was attempting to copy a non-existant
'lib' directory.
- Updated the associated CI test to better report failures.