NAS-137970 / 26.04 / fix zfs.dataset.unload_key regression (#17358)
A regression was introduced in
https://github.com/truenas/middleware/pull/17338 because the legacy
`py-libzfs` module's unmount operation was recursive by default. The
`recursive` option was missed here in that PR, furthermore, the default
setting for that property is set wrong.

Furthermore, libzfs does not allow unloading an encryption key and
keeping the dataset mounted so this method can be simplified greatly.
Even if the dataset isn't encrypted, this method should not raise an
error so there is no reason to do all the unnecessary checks.