NAS-137972 / 26.04 / remove zfs.dataset.unload_key (replace w/ zfs.resource.unload_key) (#17359)
This removes `zfs.dataset.unload_key` and replace it with
`zfs.resource.unload_key`. There was only one call site for this method
but the benefit is that the new method doesn't use our process pool. Did
some refactoring since this new method reused a lot of the same logic
from the new mount/unmount endpoints.