NAS-137669 / 26.04 / Handle `Secret` serialization for pydantic 2.10.6 (#17362)
A change in pydantic 2.10.2 makes `pydantic.Secret` a serializable type.
This is incompatible with the way that we currently serialize `Secret`
which relies on normal serialization failing and calling our custom
fallback function. Instead, manually set `pydantic.Secret`'s serialize
function.

This will allow us to upgrade to pydantic 2.10.6, the version of the
package present on trixie.

http://jenkins.eng.ixsystems.net:8080/job/tests/job/api_tests/6245/