NAS-138057 / 25.04.2.5 / Allow easier debugging of kerberos and winbind errors (by anodos325) (#17398)
It was observed on a customer system that winbindd could take some time
to allow wbclient connections after a restart during the post-join tasks
for active directory. This was observed in 25.04.2.4. This commit has
the post-join operations retry generating a libwbclient context after an
initial failure with the type that winbindd is not available. If it
still fails then rather than rolling back the current state of the
domain join, we will leave it intentially in a faulted state to allow
easier developer debugging.

Original PR: https://github.com/truenas/middleware/pull/17395

Co-authored-by: Andrew Walker <awalker@ixsystems.com>