Fix crash on AF_UNIX auth for non-root non-interactive session
It appears that in some cases scripts running via cron are not
getting a proper loginuid set and so are treated as non-interactive
sessions. There was a bug in this particular code-path where we
were accidentally setting the user_info to None.

(cherry picked from commit 661a1714ce769f2d7c4f27bd91fd4dabc7c44689)