NAS-138081 / 25.10.0 / Fix crash on AF_UNIX auth for non-root non-interactive session (by anodos325) (#17406)
It appears that in some cases scripts running via cron are not getting a
proper loginuid set and so are treated as non-interactive sessions.
There was a bug in this particular code-path where we were accidentally
setting the user_info to None.

Original PR: https://github.com/truenas/middleware/pull/17404

Co-authored-by: Andrew Walker <andrew.walker@truenas.com>