NAS-137769 / 26.04 / Truesearch integration (#17382)
This PR adds initial support for truesearch service integration.

Truesearch indexes available directories (currently, Samba shares if
Spotlight is enabled). The index is stored in the system dataset. It is
used for Spotlight search in Samba.

Truesearch service is automatically started if there are directories to
index, and stopped if there are not. No explicit user configuration is
required.

---------

Co-authored-by: Mark Grimes <126630154+mgrimesix@users.noreply.github.com>