NAS-138167 / 26.04 / Remove obsoleted NFS exports.d test. (#17431)
The CI tests related to NFS `/etc/exports.d` are failing.
The reasons for these tests was to verify actions taken in the NFS
exports mako. After ZFS was updated to prevent taking action on NFS
exports, the mako was updated and all actions relative to
`/etc/exports.d` was removed.

This PR removes the CI tests associated with testing the NFS exports
mako file processing of `/etc/exports.d`