NAS-137930 / 26.04 / Add V-series enclosure and drive identification support (#17425)
Like the M-series, the V-100 series has 24 front-facing drive slots with
4 NVMe slots in the rear. The 24 front-facing drives are split into two
groups of 12 which are each managed by a different SES device.
Determining which SES device maps to which group of drive slots requires
us to compare the SAS addresses of the two SES devices. The V-series is
unique in this sense because its SES enclosures cannot independently
determine their own drive mappings without looking at the other SES
device, requiring that we change how `Enclosure` objects are initialized
in `get_ses_enclosures`.

An alternative approach was suggested in which we wait to determine slot
mappings in `enclosure2.query`, specifically in the `combine_enclosures`
utility method, in an attempt to narrow the code changes. This presents
its own challenges however, since `Enclosure._parse_elements` must be
executed after determining slot mapping and by this time in
`enclosure2.query` we no longer have access to the `Enclosure` objects.

Passes manual slot identification testing on V160 ✅