NAS-138075 / 25.10.1 / Remove limitation on the number of NFS share host and/or network entries (by mgrimesix) (#17443)
Automatic cherry-pick failed. Please resolve conflicts by running:

    git reset --hard HEAD~1
    git cherry-pick -x 52b820a4553bf01a4982f6aecd483aa3a9b80039
    git cherry-pick -x 1cd952c1cb38cbe4327f2b4bfe44b856c094c489

If the original PR was merged via a squash, you can just cherry-pick the
squashed commit:

    git reset --hard HEAD~1
    git cherry-pick -x 30ad6a4e5dd0329fabbbf5e72359f4d25c0fca45

Removed the limitation on the number of host and network entries that
can be assigned to an NFS share.

With this change the only limitation is the size of the command buffer.
I've been able to test a summation of over 6000 of host and network
entries. At this extreme we experience command timeout (takes over 1 min
to process). This was using only IP addresses. The setting _does_ take
(despite timeout). Clients can mount and the NFS server on TrueNAS is
functional.

I expect using host names (instead of IP addresses) may introduce other
areas of slowness due to the name lookups. This is still under test to
quantify the impact.

The hard limit has been replaced with an alert. If there are 100 or more
entries in either list an alert is generated. Alerts are automatically
cleared if the list is reduced to under 100 or if the share is deleted.

CI tests have been updated/created.

Original PR: https://github.com/truenas/middleware/pull/17436

---------

Co-authored-by: Mark Grimes <126630154+mgrimesix@users.noreply.github.com>
Co-authored-by: Mark Grimes <mark.grimes@ixsystems.com>
Co-authored-by: caleb <yocalebo@gmail.com>