NAS-138225 / 26.04 / add zfs.resource.destroy (#17459)
This adds zfs.resource.destroy which makes use of our new libzfs C
module. Nothing is using this (yet) but adds the public and internal
endpoints for future consumer's consumption.

This takes two approaches for destroying resources. If a non-recursive
destructive operation is requested, it uses a non-atomic method for
destroying the resource. If a recursive destructive operation is
requested, we will use the channel program functionality. The reason why
we use the program functionality is because it's atomic and the most
"proper" way to accomplish the task.

A future PR will begin converting all internal consumers to calling this
instead of zfs.dataset.delete.

All tests pass.