NAS-138236 / 25.10.1 / Fix crash in `disk.sync_all` when a new group of disks with non-unique identifiers is presented (by themylogin) (#17472)
When a new group of disks with non-unique identifiers is presented (i.e.
two new USB disks with same `serial_lunid`, like in the test), the first
disk is inserted to the database, but it is not appended to the `qs`
array which is an in-memory list of all the disks present in the
database. So, when the next disk is processed, `qs` array does not
contain the row for the first disk, and the next disk is also treated as
a new one. We try to insert it and fail with the unique constraint
error.

We don't really care about systems that have disks with non-unique
identifiers. Let's just fix this crash and only insert one disk row.

Original PR: https://github.com/truenas/middleware/pull/17469

Co-authored-by: themylogin <themylogin@gmail.com>