NAS-138267 / 26.04 / Require pagination for audit queries (#17471)
This commit changes validation for the frontend API audit.query and
backend API for audit databases to require either querying the row count
of the audit database or specify a limit on the total rows queried.

The `limit` query-option is now restricted to less than 10K rows to
avoid potential for an API user to cause server to go OOM by setting it
to an unreasonably high value for large audit databases.

The validation for the `services` key now restricts audit queries to a
single database. This is because implementing pagination and sorting
across multiple databases is not practical without loading excessive
amounts of rows in memory.

Due to the above changes, the utilities related to parsing audit query
filters are vastly simplified, but have reduced functionality. For
example, queries filtering based on fields in event_data will no longer
be supported.

This commit breaks audit report generation, which will be fixed in
followup PR to address NAS-138241.