NAS-138291 / 26.04 / fix new container API (and tests) (#17486)
This fixes many small errors that broke the container API tests.

1. `devices` key in API schema doesn't exist in db
2. `pool.dataset.delete` is public API and so it crashed with ENOENT
when trying to delete the container dataset because we consider it a
"private" dataset
3. the container dataset has a different mountpoint that we have to
handle specially. This ended up exposing another bug in our new libzfs
library but a work-around has been implemented in the meantime (with
appropriate developer comment)
4. the tests were, yet again, calling `pool.dataset.get_instance` which
is only meant for datasets that we don't consider "private"
5. the tests were failing because of the changes pushed and fixed here:
https://github.com/truenas/middleware/pull/17485

With all these changes the container API tests are now passing with the
exception of 1. The one that is failing looks like a bad test and/or
another unrelated bug.