NAS-138294 / 26.04 / Do not import API when `middlewared.utils` is imported (#17489)
`filter_list` now is a part of the middleware API. Historically, it was
a part of `middlewared.utils` module, and so it is still imported there.

As a consequence, when a standalone script does `from
middlewared.utils.xxx import yyy`, the whole `middlewared.utils` module
is imported, and so the whole middleware API is imported. This will
increase standalone scripts startup time.

The solution is to move `filter_list` to its special
`middlewared.utils.filter_list`, so that it is only imported explicitly.