NAS-138211 / 26.04 / Add attachment delegate for containers (#17491)
## Context

We would like to have filesytem attachment delegate in place for
containers so operations like pool export/import or dataset lock/unlock
etc are gracefully handled.

However there is a unique configurational flow with containers. In
attachment delegates we query for attachments which are in some X state
so we can possibly shift it's state from X to Y. This scenario has been
handled by `ContainerFSAttachmentDelegate`. However if for example all
containers are in a stopped state and the pool is to be exported, we
would not be letting users know in any way that there are containers on
this pool (even if they are stopped).

How this works with apps is that apps has docker service running which
has it's own attachment delegate, so even if there are apps which are
not in some X state, that delegate is responsible for letting user know
that this pool has apps etc.

To improve visibility here for the user, another attachment delegate has
been added i.e `LXCFSAttachmentDelegate` which is just responsible for
letting user know that this X dataset/pool has attachments, so be
mindful as you proceed with the operation in question.

There was another case where if the pool had exported, container would
fail because `pylibvirt_container` would not handle that gracefully. For
that specific case, it is only relevant when starting the container that
we ensure that root dataset of the container does indeed actually exist
and let's fail at that point then.