NAS-138052 / 26.04 / Identify SED based disks (#17493)
This PR adds changes so we are able to better support SED based disks
for enterprise systems.
How this is designed to work is that on enterprise systems, we will mark
all disks as either SED based or not SED based. This happens once when
disks are synced and the information gets stored in the database. After
that, `disk.query` should now get us more relevant information about the
current status of SED disks if/as requested by consumer.

For the disks which we have marked as SED on enterprise systems,
unlock/setup/reset sed disk endpoints have been exposed to manage SED
disks.