NAS-138165 / 26.04 / Properly manage TPM device state (by Qubad786) (#17563)
Automatic cherry-pick failed. Please resolve conflicts by running:

    git reset --hard HEAD~1
    git cherry-pick -x 360cebe19a299a426d3ed53cbfb042b7b75eb7d4

If the original PR was merged via a squash, you can just cherry-pick the
squashed commit:

    git reset --hard HEAD~1
    git cherry-pick -x ee9d87e04a41c441e210227d6593008270e6994d

## Context

Currently when TPM is enabled, it's state is lost when guest is power
cycled. We would have to maintain it's state so that it does not get
lost. If we don't do this, the OS in this case detects TPM has changed
when guest is power cycled and prompts the user to reset credentials.

The best place to put this state would be on system dataset along with
NVRAM files so that these persist across upgrades and across HA
controllers.

Original PR: https://github.com/truenas/middleware/pull/17545

Co-authored-by: M. Rehan <mrehanlm93@gmail.com>