NAS-138491 / 26.04 / Update test for trixie changes and improve robustness. (#17584)
This test does end-to-end testing the audit subsystem using `privileged`
and `escalation` events. We used `ping` as an example of a `privileged`
command event.

The update to trixie changed the auto-generated list of privileged
commands. One was `ping`, which is no longer an audited `privileged`
command.

This PR:

- Changes the tested privilege command from `ping` to `mount`
- Update test to match recent changes in the TrueNAS audit subsystem,
specifically around filtering.
- Other changes to improve robustness and small amount of refactor.

CI was tested using a jenkins VM.