NAS-138165 / 25.10.1 / Properly manage TPM device state (#17545)
## Context

Currently when TPM is enabled, it's state is lost when guest is power
cycled. We would have to maintain it's state so that it does not get
lost. If we don't do this, the OS in this case detects TPM has changed
when guest is power cycled and prompts the user to reset credentials.

The best place to put this state would be on system dataset along with
NVRAM files so that these persist across upgrades and across HA
controllers.