NAS-138051 / 26.04 / Automatically handle SED disks during pool create/update operations (#17494)
This PR adds changes to automatically setup/unlock SED disks during pool
create/update operations.
During pool create, this only happens exclusively if `all_sed` boolean
flag is set assuming the system is licensed to use SED disks.
If this flag is not set during pool creation, we do not automatically
setup/unlock SED disks and they are treated as normal disks.

For pool update, we have not done anything special on this end and if
system is licensed for SED and SED disks are being added and need to be
unlocked/setup - that would be done as required.