NAS-138533 / 26.04 / Do not allow changing device type (#17601)
## Problem

Currently there is no validation in preventing a consumer from changing
a device type of an existing container/vm device. This can cause to
configuration issues because `attributes` is a dict and if this happens,
old attrs are not unset.

## Solution

We would like to prevent this from happening so validation has been
added which makes sure a consumer is not able to change device type.