NAS-138053 / 26.04 / Improve status detail reporting in case of failed pool import with SED disks (#17625)
This PR adds changes to improve pool error reporting in case we have had
a pool which failed to import and we had SED disks which were locked. In
this case `status_detail` attr which was `nulll ` previously in such
cases has been updated to reflect the error message showing what locked
disks we have atm.