NAS-138520 / 26.04 / Fix a few audit tests on nightlies (#17596)
- `midclt call audit.query '{"query-filters": [["event_data.params",
"=", [1]]], "query-options": {"limit": 1}}'` fails due to the
combination of using a JSON path key and a list as its query value.
Handle this case by dumping the value to a string before comparing it to
the return of JSON_EXTRACT which always returns a string.

- `misc_methods` should list methods with their arguments unpacked like
how they are passed to `Middleware.call`.

- test_root_session_alert.py: We just need to pass "limit" as per #17471

http://jenkins.eng.ixsystems.net:8080/job/tests/job/api_tests/6628/