NAS-138635 / 26.04 / Update static routes validation to support auditing. (by mgrimesix) (#17670)
### A forward port with fixup
This is preparation for adding auditing support for static route CRUD.

The static route API uses pydantic IP address types which, when included
in auditing, generate JSON serialization errors.

The fix is to use strings in the pydantic API and add the validation to
the static route CRUD handlers.

---------------------------------
**This has been manually tested with `test_staticroutes.py` CI test on a
halfmoon VM updated with this branch.**

Original PR: https://github.com/truenas/middleware/pull/17669

---------

Co-authored-by: Mark Grimes <mark.grimes@ixsystems.com>