NAS-138402 / 25.10.1 / Add FCP_SHARE SMB share purpose (by anodos325) (#17679)
This commit adds a new SMB share purpose explicilty for Final Cut Pro
per internal request as a response to a certain user demographic that
expressed displeasure in us removing the ability to override the vfs
objects (which would break our SMB server configuration in various
subtle ways). The share type does not provide novel capabilities, but
does make it slightly easier to configure an SMB share per the Apple
support document provided at

https://support.apple.com/en-ca/101919

This support document can be distilled to the following:

* vfs_fruit is required (this is the global setting in the SMB server
configuration).

The Apple support document also status that vfs_catia is required, but
the specified configuration in the document is effectively a no-op in
Samba (including version 4.3.4 which is specifed in the document). This
changeset reflects my guess that the intent of the support document was
that it requires the `native` fruit encoding for files, and so this
share type *forces* native encoding and vfs_fruit (which could have also
been set by the user simply checking a UI checkbox for "apple-style name
mangling").

Original PR: https://github.com/truenas/middleware/pull/17676

Co-authored-by: Andrew Walker <awalker@ixsystems.com>