Drop privileges to READONLY_ADMIN while generating debugs
This commit drops the effective permissions of a child process
that generates the ixdiagnose debugs so that it contains only
READONLY_ADMIN privileges. This ensures that API responses will
always be redacted. The privilege drop is *only* for the middleware
sessions created during ixdiagnose operations. Shell commands will
still be executed as root in order to bypass permissions
on configuration files and logs we may need to gather for the debug.