NAS-138680 / 26.04 / Fix NFS share alert test (#17703)
The temporary dataset and share were re-using names of existing datasets
and shares causing the test to fail.
This also occurs in Goldeye.
The test fails only when run with the other tests in the test_300_nfs.py
module.

This PR makes the names unique and allows the test to pass.

Tested locally.