Skip IPA SMB configuration if we don't have keytab
Originally we hard-coded the ipa SMB keytab path into the SMB server
configuration if TrueNAS is joined to an IPA domain. It's in principle
possible for auto-setup of SMB keytab / SPN to fail when joining an IPA
domain. In this case we will detect the IPA domain as having potential
for SMB access, but not actually be able to provide SMB server support.

(cherry picked from commit 76e1d3f9c66cdecf5d21887cb45c292dd2202ac2)