NAS-138684 / 26.04 / Skip IPA SMB configuration if we don't have keytab (#17706)
Originally we hard-coded the ipa SMB keytab path into the SMB server
configuration if TrueNAS is joined to an IPA domain. It's in principle
possible for auto-setup of SMB keytab / SPN to fail when joining an IPA
domain. In this case we will detect the IPA domain as having potential
for SMB access, but not actually be able to provide SMB server support.